Move /rules to /services/rules
