+++ x-pack/plugins/observability/public/pages/alerts/components/alert_actions.test.tsx
@@ -9,9 +9,9 @@ import { kibanaStartMock } from '../../../utils/kibana_react.mock';
 import React from 'react';
 import { mountWithIntl, nextTick } from '@kbn/test-jest-helpers';
 import { AlertActions, Props } from './alert_actions';
-import { inventoryThresholdAlert } from '../../../rules/fixtures/example_alerts';
+import { inventoryThresholdAlert } from '../../../services/rules/fixtures/example_alerts';
 import { RULE_DETAILS_PAGE_ID } from '../../rule_details/constants';
-import { createObservabilityRuleTypeRegistryMock } from '../../../rules/observability_rule_type_registry_mock';
+import { createObservabilityRuleTypeRegistryMock } from '../../../services/rules/create_observability_rule_type_registry_mock';
 import { TimelineNonEcsData } from '@kbn/timelines-plugin/common';
 import * as pluginContext from '../../../hooks/use_plugin_context';
 import { ConfigSchema, ObservabilityPublicPluginsStart } from '../../../plugin';
