+++ x-pack/plugins/observability/public/pages/alerts/helpers/parse_alert.ts
@@ -18,7 +18,7 @@ import { legacyExperimentalFieldMap } from '@kbn/alerts-as-data-utils';
 import { parseTechnicalFields } from '@kbn/rule-registry-plugin/common/parse_technical_fields';
 import { parseExperimentalFields } from '@kbn/rule-registry-plugin/common/parse_experimental_fields';
 import { asDuration, asPercent } from '../../../../common/utils/formatters';
-import { ObservabilityRuleTypeRegistry } from '../../../rules/create_observability_rule_type_registry';
+import { ObservabilityRuleTypeRegistry } from '../../../services/rules/create_observability_rule_type_registry';
 import type { TopAlert } from '../../../types/alerts';
 
 export const parseAlert =
