+++ x-pack/plugins/observability/public/pages/rules/rules.test.tsx
@@ -13,7 +13,7 @@ import { RulesPage } from './rules';
 import { kibanaStartMock } from '../../utils/kibana_react.mock';
 import * as pluginContext from '../../hooks/use_plugin_context';
 import { KibanaPageTemplate } from '@kbn/shared-ux-page-kibana-template';
-import { createObservabilityRuleTypeRegistryMock } from '../../rules/observability_rule_type_registry_mock';
+import { createObservabilityRuleTypeRegistryMock } from '../../services/rules/create_observability_rule_type_registry_mock';
 import { AppMountParameters } from '@kbn/core/public';
 import { ALERTS_FEATURE_ID } from '@kbn/alerting-plugin/common';
