+++ x-pack/plugins/observability/public/services/rules/create_observability_rule_type_registry.ts
@@ -11,7 +11,7 @@ import {
   RuleTypeRegistryContract,
 } from '@kbn/triggers-actions-ui-plugin/public';
 import { ParsedTechnicalFields } from '@kbn/rule-registry-plugin/common/parse_technical_fields';
-import { AsDuration, AsPercent } from '../../common/utils/formatters';
+import { AsDuration, AsPercent } from '../../../common/utils/formatters';
 
 export type ObservabilityRuleTypeFormatter = (options: {
   fields: ParsedTechnicalFields & Record<string, any>;
