+++ x-pack/plugins/observability/public/services/rules/create_observability_rule_type_registry_mock.ts
