+++ x-pack/plugins/observability/public/services/rules/fixtures/example_alerts.ts
