+++ x-pack/plugins/observability/public/services/rules/register_observability_rule_types.ts
@@ -8,11 +8,11 @@ import { lazy } from 'react';
 import { i18n } from '@kbn/i18n';
 import { ALERT_REASON } from '@kbn/rule-data-utils';
 
-import { SLO_ID_FIELD } from '../../common/field_names/infra_metrics';
-import { ConfigSchema } from '../plugin';
+import { SLO_ID_FIELD } from '../../../common/field_names/infra_metrics';
+import { ConfigSchema } from '../../plugin';
 import { ObservabilityRuleTypeRegistry } from './create_observability_rule_type_registry';
-import { SLO_BURN_RATE_RULE_ID } from '../../common/constants';
-import { validateBurnRateRule } from '../components/burn_rate_rule_editor/validation';
+import { SLO_BURN_RATE_RULE_ID } from '../../../common/constants';
+import { validateBurnRateRule } from '../../components/burn_rate_rule_editor/validation';
 
 export const registerObservabilityRuleTypes = (
   config: ConfigSchema,
@@ -33,7 +33,7 @@ export const registerObservabilityRuleTypes = (
     documentationUrl(docLinks) {
       return 'https://www.elastic.co/guide/en/observability/current/slo-burn-rate-alert.html';
     },
-    ruleParamsExpression: lazy(() => import('../components/burn_rate_rule_editor')),
+    ruleParamsExpression: lazy(() => import('../../components/burn_rate_rule_editor')),
     validate: validateBurnRateRule,
     requiresAppContext: false,
     defaultActionMessage: i18n.translate(
