+++ x-pack/plugins/observability/public/utils/kibana_react.storybook_decorator.tsx
@@ -14,7 +14,7 @@ import { CoreTheme } from '@kbn/core-theme-browser';
 import { MemoryRouter } from 'react-router-dom';
 import { casesFeatureId, sloFeatureId } from '../../common';
 import { PluginContext } from '../context/plugin_context/plugin_context';
-import { createObservabilityRuleTypeRegistryMock } from '../rules/observability_rule_type_registry_mock';
+import { createObservabilityRuleTypeRegistryMock } from '../services/rules/create_observability_rule_type_registry_mock';
 import { ConfigSchema } from '../plugin';
 
 export function KibanaReactStorybookDecorator(Story: ComponentType) {
